app/requirements.txt
Flask==1.0.2
Flask-PyMongo==2.2.0
requests==2.20.1
python-slugify==4.0.0